requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
redis[hiredis]==5.1.1
aiokafka==0.11.0
pybreaker==1.2.0
python-dotenv==1.0.1
httpx==0.27.2
python-dateutil==2.9.0
email-validator==2.2.0

